Build with Meson
Has the following meson build options:
* default-wallpaper
* plank-dockitems
* default-gsettings-overrides

All default to true.

We also split the default-settings gsettings override into another that
only sets plank's default dockitems. This is installed when plank-dockitems
is true. We need to have this because those settings are dependent on
those dockitems actually being installed and we have it optional.